chore: move Items-related components and hooks into pages/Items directory
