+++ src/pages/Items/components/AllProducts/AllProducts.js
